requirements.txt
Django>=5.2,<6
djangorestframework>=3.14,<4
django-cors-headers>=4.2,<5
mysqlclient>=2.1,<3
drf-yasg>=1.21,<2
celery>=5.3,<6
pika>=1.3,<2
python-dotenv>=1.1,<2
django-environ==0.12.0
